get apps that can CONNECT to support default vs. configed responses
Get all of the responders that are able to make a full connection using
Panic.ACTION_CONNECT, which is used to configure the response an
app makes.  For destructive responses, it is essential that the trigger
and responder are connected in order to prevent random apps from making
responders destroy things.  Apps can also only respond to
Panic.ACTION_TRIGGER with a non-destructive response which does
not require the apps to connect or the user to configure anything.