PanicTrigger.sendTrigger() works from Receivers and Services
A Service or BroadcastReceiver often does not have an associated Activity,
so forcing triggers to be sent only with an Activity makes a number of use
cases quite difficult.  This is a temporary workaround, it just sends using
Context.startActivity() if an Activity was not provided.  That means that
the receiver will not be able to verify which app sent that trigger Intent
when the trigger is sent from a Service or BroadcastReceiver.